% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_map.R
\name{inat_map}
\alias{inat_map}
\title{Plot iNaturalist observations}
\usage{
inat_map(data)
}
\arguments{
\item{data}{data frame of iNaturalist observations}
}
\value{
a leaflet map object
}
\description{
Plot observations from iNaturalist for Mexico.
}
\examples{
\dontrun{
  m_obs <- get_inat_obs(taxon="Ambystoma maculatum")
  salamander_map <- inat_map(m_obs,plot=FALSE)
  ### Now we can modify the returned map
  salamander_map + borders("state") + theme_bw()
}
}
